+++ sync/track.h
@@ -26,7 +26,7 @@ struct track_key {
 struct sync_track {
 	char *name;
 	struct track_key *keys;
-	size_t num_keys;
+	int num_keys;
 };
 
 int sync_find_key(const struct sync_track *, int);
@@ -41,7 +41,7 @@ static inline int key_idx_floor(const struct sync_track *t, int row)
 #ifndef SYNC_PLAYER
 int sync_set_key(struct sync_track *, const struct track_key *);
 int sync_del_key(struct sync_track *, int);
-static inline int is_key_frame(const struct sync_track *t, size_t row)
+static inline int is_key_frame(const struct sync_track *t, int row)
 {
 	return sync_find_key(t, row) >= 0;
 }
